removed wewather api and retreive the weather data from backemd
