+++ components/Screens/Main/Main.tsx
@@ -50,7 +50,7 @@ const Main = (): ReactElement => {
           </div>
         </S.ResultSection>
         <S.WeatherSection>
-          {/* <WeatherCard deviceType={deviceType} /> */}
+          <WeatherCard deviceType={deviceType} />
         </S.WeatherSection>
       </S.MainLayout>
     </S.Wrapper>
